+++ shared/src/api/client/api/windows.ts
@@ -1,4 +1,4 @@
-import { Observable, Subscription } from 'rxjs'
+import { Observable, Subject, Subscription } from 'rxjs'
 import * as sourcegraph from 'sourcegraph'
 import { createProxyAndHandleRequests } from '../../common/proxy'
 import { ExtWindowsAPI } from '../../extension/api/windows'
@@ -18,6 +18,9 @@ export interface ClientWindowsAPI {
     $showNotification(message: string): void
     $showMessage(message: string): Promise<void>
     $showInputBox(options?: sourcegraph.InputBoxOptions): Promise<string | undefined>
+    $startProgress(options: sourcegraph.ProgressOptions): Promise<number>
+    $updateProgress(handle: number, progress: sourcegraph.Progress): void
+    $errorProgress(handle: number, error: any): void
 }
 
 /** @internal */
@@ -40,7 +43,8 @@ export class ClientWindows implements ClientWindowsAPI {
          * Called when the client receives a window/showInput request and expected to return a promise that
          * resolves to the user's input.
          */
-        private showInput: (params: ShowInputParams) => Promise<string | null>
+        private showInput: (params: ShowInputParams) => Promise<string | null>,
+        private createProgressReporter: (options: sourcegraph.ProgressOptions) => Subject<sourcegraph.Progress>
     ) {
         this.proxy = createProxyAndHandleRequests('windows', connection, this)
 
@@ -88,6 +92,37 @@ export class ClientWindows implements ClientWindowsAPI {
                 // external API.
                 v === null ? undefined : v
         )
+    }
+
+    private handles = 1
+    private progress = new Map<number, Subject<sourcegraph.Progress>>()
+
+    public async $startProgress(options: sourcegraph.ProgressOptions): Promise<number> {
+        const handle = this.handles++
+        const progress = this.createProgressReporter(options)
+        this.progress.set(handle, progress)
+        return handle
+    }
+
+    public $updateProgress(handle: number, progress: sourcegraph.Progress): void {
+        const reporter = this.progress.get(handle)
+        if (!reporter) {
+            console.warn('No ProgressReporter for handle ' + handle)
+            return
+        }
+        reporter.next(progress)
+        if (progress.percentage && progress.percentage >= 100) {
+            reporter.complete()
+        }
+    }
+
+    public $errorProgress(handle: number, error: any): void {
+        const reporter = this.progress.get(handle)
+        if (!reporter) {
+            console.warn('No ProgressReporter for handle ' + handle)
+            return
+        }
+        reporter.error(error)
     }
 
     public unsubscribe(): void {