+++ shared/src/api/extension/api/windows.ts
@@ -1,4 +1,6 @@
+import { Observer } from 'rxjs'
 import * as sourcegraph from 'sourcegraph'
+import { asError } from '../../../util/errors'
 import { ClientCodeEditorAPI } from '../../client/api/codeEditor'
 import { ClientWindowsAPI } from '../../client/api/windows'
 import { ViewComponentData } from '../../client/model'
@@ -35,6 +37,38 @@ class ExtWindow implements sourcegraph.Window {
 
     public showInputBox(options?: sourcegraph.InputBoxOptions): Promise<string | undefined> {
         return this.windowsProxy.$showInputBox(options)
+    }
+
+    public async withProgress<R>(
+        options: sourcegraph.ProgressOptions,
+        task: (reporter: sourcegraph.ProgressReporter) => Promise<R>
+    ): Promise<R> {
+        const reporter = await this.showProgress(options)
+        try {
+            const result = await task(reporter)
+            reporter.complete()
+            return result
+        } catch (err) {
+            reporter.error(err)
+            throw err
+        }
+    }
+
+    public async showProgress(options: sourcegraph.ProgressOptions): Promise<sourcegraph.ProgressReporter> {
+        const handle = await this.windowsProxy.$startProgress(options)
+        const reporter: Observer<sourcegraph.Progress> = {
+            next: (progress: sourcegraph.Progress): void => {
+                this.windowsProxy.$updateProgress(handle, progress)
+            },
+            error: (err: any): void => {
+                const error = asError(err)
+                this.windowsProxy.$errorProgress(handle, { ...error, message: error.message, stack: error.stack })
+            },
+            complete: (): void => {
+                this.windowsProxy.$updateProgress(handle, { percentage: 100 })
+            },
+        }
+        return reporter
     }
 
     public toJSON(): any {